scripts: share string-aware Lean comment stripping
